fixes portfolio to handle floats.

+++ models/portfolio.go
@@ -1,12 +1,12 @@
 package models
 
 type Portfolio struct {
-	AvailableFunds           int    `json:"available_funds"`
-	AvailableWithdrawalFunds int    `json:"available_withdrawal_funds"`
-	Balance                  int    `json:"balance"`
+	AvailableFunds           float64    `json:"available_funds"`
+	AvailableWithdrawalFunds float64    `json:"available_withdrawal_funds"`
+	Balance                  float64    `json:"balance"`
 	Currency                 string `json:"currency"`
-	Equity                   int    `json:"equity"`
-	InitialMargin            int    `json:"initial_margin"`
-	MaintenanceMargin        int    `json:"maintenance_margin"`
-	MarginBalance            int    `json:"margin_balance"`
+	Equity                   float64    `json:"equity"`
+	InitialMargin            float64    `json:"initial_margin"`
+	MaintenanceMargin        float64    `json:"maintenance_margin"`
+	MarginBalance            float64    `json:"margin_balance"`
 }
